feat: IssueCouponRestTemplateClient 구현 및 예외 처리 추가
